now on typescript
